one class per file, minor cleanups
