+++ .fvmrc
@@ -1,3 +1,3 @@
 {
-  "flutter": "3.24.4"
+  "flutter": "3.27.3"
 }
