+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "dart.flutterSdkPath": ".fvm/versions/3.27.3"
+}
