+++ lib/coins/abstract/coin.dart
@@ -1,0 +1,32 @@
+import 'package:cupcake/coins/abstract/coin_strings.dart';
+import 'package:cupcake/coins/abstract/coin_wallet.dart';
+import 'package:cupcake/coins/abstract/coin_wallet_info.dart';
+import 'package:cupcake/coins/types.dart';
+
+enum Coins { monero, unknown }
+
+abstract class Coin {
+  Coins get type => Coins.unknown;
+
+  CoinStrings get strings;
+
+  bool get isEnabled;
+
+  Future<List<CoinWalletInfo>> get coinWallets;
+
+  Future<CoinWallet> createNewWallet(
+    String walletName,
+    String walletPassword, {
+    ProgressCallback? progressCallback,
+    required bool? createWallet,
+    required String? seed,
+    required int? restoreHeight,
+    required String? primaryAddress,
+    required String? viewKey,
+    required String? spendKey,
+    required String? seedOffsetOrEncryption,
+  });
+
+  Future<CoinWallet> openWallet(CoinWalletInfo walletInfo,
+      {required String password});
+}
