+++ lib/coins/abstract/coin_exception.dart
@@ -1,0 +1,11 @@
+class CoinException implements Exception {
+  CoinException(this.exception, {this.details});
+
+  String exception;
+  String? details;
+
+  @override
+  String toString() {
+    return "$exception\n$details";
+  }
+}
