+++ lib/coins/abstract/wallet_seed_detail.dart
@@ -1,0 +1,13 @@
+import 'package:cupcake/coins/types.dart';
+
+class WalletSeedDetail {
+  WalletSeedDetail({
+    required this.type,
+    required this.name,
+    required this.value,
+  });
+
+  final WalletSeedDetailType type;
+  final String name;
+  final String value;
+}
