+++ lib/coins/list.dart
@@ -1,4 +1,4 @@
-import 'package:cupcake/coins/abstract.dart';
+import 'package:cupcake/coins/abstract/coin.dart';
 import 'package:cupcake/coins/monero/coin.dart';
 
 const moneroEnabled = bool.fromEnvironment("COIN_MONERO", defaultValue: true);
