+++ lib/coins/types.dart
@@ -1,0 +1,6 @@
+typedef ProgressCallback = int Function({String? title, String? description});
+
+enum WalletSeedDetailType {
+  text,
+  qr,
+}
