+++ lib/utils/alerts/widget.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+Future<void> showAlertWidget({
+  required BuildContext context,
+  required String title,
+  required List<Widget> body,
+  String ok = "ok",
+}) async {
+  return showDialog<void>(
+    context: context,
+    barrierDismissible: false, // user must tap button!
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text(title),
+        content: SingleChildScrollView(
+          child: ListBody(
+            children: body,
+          ),
+        ),
+        actions: <Widget>[
+          TextButton(
+            child: Text(ok),
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+          ),
+        ],
+      );
+    },
+  );
+}
