+++ lib/utils/form/abstract_form_element.dart
@@ -1,0 +1,5 @@
+abstract class FormElement {
+  bool get isOk => true;
+  String get label => throw UnimplementedError();
+  Future<String> get value => throw UnimplementedError();
+}
