+++ lib/utils/form/abstract_value_outcome.dart
@@ -1,0 +1,6 @@
+abstract class ValueOutcome {
+  Future<void> encode(String input);
+  Future<String> decode(String output);
+
+  String get uniqueId => throw UnimplementedError();
+}
