+++ lib/utils/form/default_validator.dart
@@ -1,0 +1,3 @@
+String? defaultFormValidator(String? input) {
+  return null;
+}
