+++ lib/utils/form/flutter_secure_storage_value_outcome.dart
@@ -1,0 +1,72 @@
+import 'dart:convert';
+import 'dart:math';
+
+import 'package:crypto/crypto.dart';
+import 'package:cupcake/utils/config.dart';
+import 'package:cupcake/utils/form/abstract_value_outcome.dart';
+import 'package:cupcake/utils/secure_storage.dart';
+
+class FlutterSecureStorageValueOutcome implements ValueOutcome {
+  FlutterSecureStorageValueOutcome(this.key,
+      {required this.canWrite, required this.verifyMatching});
+
+  final String key;
+  final bool canWrite;
+  final bool verifyMatching;
+
+  @override
+  Future<void> encode(String input) async {
+    List<int> bytes = utf8.encode(input);
+    Digest sha512Hash = sha512.convert(bytes);
+    var valInput =
+        await secureStorage.read(key: "FlutterSecureStorageValueOutcome._$key");
+    if (valInput == null) {
+      await secureStorage.write(
+          key: "FlutterSecureStorageValueOutcome._$key",
+          value: sha512Hash.toString());
+      valInput = await secureStorage.read(
+          key: "FlutterSecureStorageValueOutcome._$key");
+    }
+    if (sha512Hash.toString() != valInput && verifyMatching) {
+      throw Exception("Input doesn't match the secure element value");
+    }
+
+    final input_ = await secureStorage.read(key: key);
+    // Do not update secret if it is already set.
+    if (input_ != null) {
+      return;
+    }
+    if (!canWrite) {
+      if (config.debug) {
+        throw Exception(
+            "DEBUG_ONLY: canWrite is false but we tried to flush the value");
+      }
+      return;
+    }
+    var random = Random.secure();
+    var values = List<int>.generate(64, (i) => random.nextInt(256));
+    final pass = base64Url.encode(values);
+    await secureStorage.write(key: key, value: pass);
+    return;
+  }
+
+  @override
+  Future<String> decode(String output) async {
+    List<int> bytes = utf8.encode(output);
+    Digest sha512Hash = sha512.convert(bytes);
+    var valInput =
+        await secureStorage.read(key: "FlutterSecureStorageValueOutcome._$key");
+    if (sha512Hash.toString() != valInput && verifyMatching) {
+      throw Exception("Input doesn't match the secure element value");
+    }
+    final input = await secureStorage.read(key: key);
+    if (input == null) {
+      throw Exception("no secure storage $key found");
+    }
+    return "$input/$output";
+  }
+
+  @override
+  // TODO: implement uniqueId
+  String get uniqueId => key;
+}
