+++ lib/view_model/open_wallet_view_model.dart
@@ -1,7 +1,8 @@
-import 'package:cupcake/coins/abstract.dart';
+import 'package:cupcake/coins/abstract/coin_wallet_info.dart';
 import 'package:cupcake/utils/call_throwable.dart';
+import 'package:cupcake/utils/form/flutter_secure_storage_value_outcome.dart';
+import 'package:cupcake/utils/form/pin_form_element.dart';
 import 'package:cupcake/view_model/abstract.dart';
-import 'package:cupcake/view_model/create_wallet_view_model.dart';
 import 'package:cupcake/views/wallet_home.dart';
 import 'package:flutter/cupertino.dart';
