+++ lib/view_model/unconfirmed_transaction_view_model.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:cupcake/coins/abstract.dart';
+import 'package:cupcake/coins/abstract/coin_wallet.dart';
 import 'package:cupcake/view_model/abstract.dart';
 import 'package:flutter/cupertino.dart';
