+++ lib/view_model/wallet_edit_view_model.dart
@@ -1,5 +1,8 @@
-import 'package:cupcake/coins/abstract.dart';
-import 'package:cupcake/view_model/create_wallet_view_model.dart';
+import 'package:cupcake/coins/abstract/coin_wallet_info.dart';
+import 'package:cupcake/utils/form/abstract_form_element.dart';
+import 'package:cupcake/utils/form/flutter_secure_storage_value_outcome.dart';
+import 'package:cupcake/utils/form/pin_form_element.dart';
+import 'package:cupcake/utils/form/string_form_element.dart';
 import 'package:flutter/material.dart';
 import 'package:cupcake/view_model/abstract.dart';
 import 'package:path/path.dart' as p;
