+++ lib/view_model/wallet_home_view_model.dart
@@ -1,4 +1,5 @@
-import 'package:cupcake/coins/abstract.dart';
+import 'package:cupcake/coins/abstract/coin.dart';
+import 'package:cupcake/coins/abstract/coin_wallet.dart';
 import 'package:cupcake/view_model/abstract.dart';
 import 'package:cupcake/views/barcode_scanner.dart';
 import 'package:flutter/cupertino.dart';
