+++ lib/views/open_wallet.dart
@@ -1,7 +1,7 @@
-import 'package:cupcake/coins/abstract.dart';
+import 'package:cupcake/coins/abstract/coin_wallet_info.dart';
 import 'package:cupcake/view_model/open_wallet_view_model.dart';
 import 'package:cupcake/views/abstract.dart';
-import 'package:cupcake/widgets/form_builder.dart';
+import 'package:cupcake/views/widgets/form_builder.dart';
 import 'package:flutter/cupertino.dart';
 
 // ignore: must_be_immutable
