+++ lib/views/receive.dart
@@ -1,4 +1,4 @@
-import 'package:cupcake/coins/abstract.dart';
+import 'package:cupcake/coins/abstract/coin_wallet.dart';
 import 'package:cupcake/view_model/receive_view_model.dart';
 import 'package:cupcake/views/abstract.dart';
 import 'package:flutter/cupertino.dart';
