+++ lib/views/settings.dart
@@ -1,6 +1,5 @@
-import 'package:cupcake/utils/alert.dart';
+import 'package:cupcake/utils/alerts/widget.dart';
 import 'package:cupcake/utils/config.dart';
-import 'package:cupcake/utils/secure_storage.dart';
 import 'package:cupcake/view_model/settings_view_model.dart';
 import 'package:cupcake/views/abstract.dart';
 import 'package:flutter/cupertino.dart';
